+++ raster/rt_core/rt_wkb.c
@@ -41,8 +41,8 @@ rt_band_from_wkb(
 	rt_band band = NULL;
 	int pixbytes = 0;
 	uint8_t type = 0;
-	unsigned long sz = 0;
-	uint32_t v = 0;
+	size_t sz = 0;
+	size_t v = 0;
 
 	assert(NULL != ptr);
 	assert(NULL != end);
@@ -195,7 +195,7 @@ rt_band_from_wkb(
 	}
 
 	/* This is an on-disk band */
-	sz = width * height * pixbytes;
+	sz = (size_t)width * height * pixbytes;
 	if (((*ptr) + sz) > end) {
 		rterror("rt_band_from_wkb: Premature end of WKB on band data reading (%s:%d)",
 			__FILE__, __LINE__);
@@ -234,7 +234,7 @@ rt_band_from_wkb(
 			}
 
 			flipme = band->data.mem;
-			sz = width * height;
+			sz = (size_t)width * height;
 			for (v = 0; v < sz; ++v) {
 				flipper(flipme);
 				flipme += pixbytes;
@@ -250,7 +250,7 @@ rt_band_from_wkb(
 		uint8_t maxVal = band->pixtype == PT_1BB ? 1 : (band->pixtype == PT_2BUI ? 3 : 15);
 		uint8_t val;
 
-		sz = width*height;
+		sz = (size_t)width*height;
 		for (v = 0; v < sz; ++v) {
 			val = ((uint8_t*) band->data.mem)[v];
 			if (val > maxVal) {
